Seperate map controller and state holder classes and add changes controller
