+++ WebSocketProjects/WebSocketServerWorking/WebSocketServerWorking/SocketComunicator.cs
@@ -14,7 +14,7 @@ namespace WebSocketServerWorking
     {
         //stuff for getting Sessions object
         public bool connectedToMap = false;
-        MapData mapData = null;
+        MapController mapController = null;
 
         protected override void OnMessage(MessageEventArgs e) 
         {
@@ -41,32 +41,32 @@ namespace WebSocketServerWorking
             message["type"] = "message";
             message["message"] = data["message"].ToString();
             message["sender"] = ServerController.GetMapData(0).GetPlayerTagByID(this.ID);
-            mapData.SendAll(message);
+            mapController.SendAll(message);
         }
 
         protected void UpdateLocationMessageReceived(JObject data)
         {
-            data["id"] = mapData.GetPlayerTagByID(this.ID)["id"];
-            mapData.UpdatePlayerLocation(data);
+            data["id"] = mapController.GetPlayerTagByID(this.ID)["id"];
+            mapController.changesController.UpdatePlayerLocation(data);
         }
 
         protected override void OnOpen()
         {
             if (!connectedToMap) // conenect sessions object to map
             {
-                mapData = ServerController.GetMapData(0);
-                connectedToMap = mapData.ConnectSessions(Sessions);
+                mapController = ServerController.GetMapData(0);
+                connectedToMap = mapController.ConnectSessions(Sessions);
             }
             string username = Context.CookieCollection["username"].Value; // TODO seperate username and nickname
-            int id = ServerController.GetMapData(0).RegisterPlayer(new Player(this.ID, username, username));
+            int id = mapController.RegisterPlayer(new Player(this.ID, username, username));
             Console.WriteLine(this.ID + ", id = " + id + " connected successfully");
-            mapData.UpdateClientsFull();
+            mapController.UpdateClientsFull();
         }
 
         protected override void OnClose(CloseEventArgs e)
         {
-            mapData.UnregisterPlayer(mapData.FindPlayer(this.ID));
-            mapData.UpdateClientsPartial();
+            mapController.UnregisterPlayer(mapController.FindPlayer(this.ID));
+            mapController.UpdateClientsPartial();
             base.OnClose(e);
         }
